[*] Rewrite alarm classes
[+] Add size check for update data

+++ widgets/wd_func.cpp
@@ -412,7 +412,7 @@ QStatusBar *WDFunc::NewSB(QWidget *w)
     msgConnectionType->setObjectName("ConnectionType");
 
     QObject::connect(Board::GetInstance(), &Board::typeChanged, [msgModel, msgSerialNumber]() {
-        quint32 serialNumber = Board::GetInstance()->type();
+        quint16 serialNumber = Board::GetInstance()->type();
         QString deviceName = QVariant::fromValue(Board::DeviceModel(serialNumber)).toString();
         msgModel->setText(deviceName);
     });
@@ -447,7 +447,7 @@ QPixmap WDFunc::NewCircle(QColor color, float radius)
     gradient.setColorAt(1, Qt::black);
     painter.setBrush(QBrush(gradient));
     painter.drawEllipse(0, 0, myPix.width() - painter.pen().width(), myPix.height() - painter.pen().width());
-    ;
+
     return myPix;
 }
